utils: move split_quoted to the string submodule
